update helpers.js - refactor randomized pairing algorithm to reduce runtime from O(infinity) to O(n^2)
Previously the randomized pairing algorithm relied on a limited minimum pairing distance to the size of the cohort, combined with the statistical certainty that by randomly selecting we will eventually stumble upon a valid pair. Because of randomness, the theoretical maximum runtime is O(infinity). By instead computing the valid edges, and using recurison to loop through all possible edge combinations, we reduce the maximum theoretical runtime to O(n^3)